backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pymongo>=4.6.0
scikit-learn>=1.4.0
sentence-transformers>=2.2.0
google-api-python-client>=2.100.0
python-dotenv>=1.0.0
certifi>=2024.0.0
